[agent] refactor(assembly): extract locale + template modules from lib.rs
